add an event listener for local_update_applied events
Currently the local_update_applied events are happing before the 'then'
part of the following promise which means that we are not able to catch
those events. This is a suggestion to add an explicit listener for such
events so that we can verify that that occur and have the expected data.