FH-3114 Add `listCollisions` and `removeCollision` to collision test
Currently the collision test is asserting that a collision can be caused.
However the functions provided by the `fh-js-sdk` `listCollisions` and
`removeCollision` are not being tested. The definition for these are:

  * listCollisions - List all records in the `_collision` collection in MongoDB.
  * removeCollision - Remove a record from a collection in MongoDB.

This updates the collision test to use these functions and ensure they are always
up to date, as they currently calling the back-end for each invocation. If this
changes the behaviour should still stay the same.