FH-3131 Add dataset initialisation endpoint to test server
Currently the default sync frequency for a dataset on both client
and server sides is ten seconds. This causes the tests to last
longer than they need to.

This adds and endpoint to the server side of the tests and invokes
it from the client side for the main `specDataset` dataset with a
`syncFrequency` of one second.

Testing it out a bit has shown the length of the tests half. So
that's always a good thing. It also allows for the removal of
custom timeouts on specific tests, which is also nice.