FH-3151 Assert getUID behaviour
`getUID` references a map that contains hashes of records with the UID
of that record in MongoDB. This is because when a record is first created
it is given a temporary UID, its hash. At this point `getUID` will not
contain the `Hash -> UID` map so it will just return back the records hash
to act as the temporary UID.

When the record is updated remotely the updated, permanent, UID will be
returned to the client. This is then added to the `Hash -> UID` map, from
this point on passing the records hash to `getUID` will return the MongoDB
UID.

This adds a test which will perform the actions above and assert that the
correct values are being returned depending on whether the remove update
has been made yet.

  * Create a record
  * Assert that `getUID` returns the records hash (temporary UID).
  * Wait for the remote update to be applied.
  * Assert that `getUID` returns the records permanent UID.