add test to verify crashed server immediate resp
Motivation:
This is work in progress as it depends on a currently unmerged pull
request:
https://github.com/feedhenry/fh-mbaas-api/pull/46

Currently, in the case where the sync server responds immediately with
no updates and stays down for at least one sync loop, the client will
mark the dataset as crashed on that sync loop, causing all data to be
resent after a number of sync loops.

Modifications:
Added a test that tries to verify this behaviour using the server in
this project.

Result:
After npm linking fh-mbaas-api (first npm link from the fh-mbaas-api
directory and then npm link fh-mbaas-api from this project) the tests
should by running npm test

JIRA:
https://issues.jboss.org/browse/FH-3132