FH-3136 Assert record is inFlight forever if server crashes
When the server is responding immediately, in the case where the
client marks a pending record as inFlight, the server responds to
the sync request but then crashes and loses all pending jobs to be
completed; the client will never resend this record as it is marked
as inFlight, leaving it as an inFlight record forever. As the record
is not marked as crashed it will never be resent.

This emulates this situation by replying to the first sync request
without Sync, meaning it does not store the pending job in memory.
This means the client never receives any updates about the record,
leaving it in an inFlight state forever.